Remove dispatch utility function from impl
