+++ tt_metal/api/tt-metalium/experimental/tensor/impl/tensor_impl.hpp
@@ -55,37 +55,4 @@ std::vector<T> to_row_major_layout(const Shape2D& shape, const Tile& tile, ttsl:
         transpose_of_faces);
 }
 
-template <typename T>
-std::vector<T> to_tile_major_layout(const Shape2D& shape, const Tile& tile, ttsl::Span<const T> data_to_convert);
-
-// Empty structs to facilitate Tensor template logic.
-struct bfloat4_b {};
-struct bfloat8_b {};
-
-// Utility to convert runtime DataType to compile-time constant and dispatch the function call
-template <typename Func, typename... Args>
-auto dispatch(DataType dtype, Func&& func, Args&&... args) {
-    switch (dtype) {
-        case DataType::BFLOAT16:
-            return (std::forward<Func>(func)).template operator()<bfloat16>(std::forward<Args>(args)...);
-        case DataType::FLOAT32:
-            return (std::forward<Func>(func)).template operator()<float>(std::forward<Args>(args)...);
-        case DataType::INT32:
-            return (std::forward<Func>(func)).template operator()<int32_t>(std::forward<Args>(args)...);
-        case DataType::UINT32:
-            return (std::forward<Func>(func)).template operator()<uint32_t>(std::forward<Args>(args)...);
-        case DataType::UINT16:
-            return (std::forward<Func>(func)).template operator()<uint16_t>(std::forward<Args>(args)...);
-        case DataType::UINT8:
-            return (std::forward<Func>(func)).template operator()<uint8_t>(std::forward<Args>(args)...);
-        case DataType::BFLOAT8_B:
-            return (std::forward<Func>(func)).template operator()<bfloat8_b>(std::forward<Args>(args)...);
-        case DataType::BFLOAT4_B:
-            return (std::forward<Func>(func)).template operator()<bfloat4_b>(std::forward<Args>(args)...);
-        case DataType::FP8_E4M3:
-            return (std::forward<Func>(func)).template operator()<float8_e4m3>(std::forward<Args>(args)...);
-        default: TT_THROW("Unsupported data type");
-    }
-}
-
 }  // namespace tt::tt_metal::tensor_impl
