+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -6,10 +6,12 @@
 #include <cstdint>
 #include <optional>
 #include <span>
+#include <utility>
 
 #include <tt-metalium/mesh_device.hpp>
 #include <tt-metalium/tilize_utils.hpp>
 #include <tt-metalium/experimental/tensor/impl/tensor_impl.hpp>
+#include <tt_stl/assert.hpp>
 
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/tensor/tensor_utils.hpp"
@@ -17,6 +19,36 @@
 #include "ttnn/tensor/layout/tensor_layout.hpp"
 
 namespace ttnn::tensor_impl {
+
+// Empty structs to facilitate Tensor template logic.
+struct bfloat4_b {};
+struct bfloat8_b {};
+
+// Utility to convert runtime DataType to compile-time constant and dispatch the function call
+template <typename Func, typename... Args>
+auto dispatch(tt::tt_metal::DataType dtype, Func&& func, Args&&... args) {
+    switch (dtype) {
+        case tt::tt_metal::DataType::BFLOAT16:
+            return (std::forward<Func>(func)).template operator()<bfloat16>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::FLOAT32:
+            return (std::forward<Func>(func)).template operator()<float>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::INT32:
+            return (std::forward<Func>(func)).template operator()<int32_t>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::UINT32:
+            return (std::forward<Func>(func)).template operator()<uint32_t>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::UINT16:
+            return (std::forward<Func>(func)).template operator()<uint16_t>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::UINT8:
+            return (std::forward<Func>(func)).template operator()<uint8_t>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::BFLOAT8_B:
+            return (std::forward<Func>(func)).template operator()<bfloat8_b>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::BFLOAT4_B:
+            return (std::forward<Func>(func)).template operator()<bfloat4_b>(std::forward<Args>(args)...);
+        case tt::tt_metal::DataType::FP8_E4M3:
+            return (std::forward<Func>(func)).template operator()<float8_e4m3>(std::forward<Args>(args)...);
+        default: TT_THROW("Unsupported data type");
+    }
+}
 
 // ===============================================================================================================================================
 //                                                              High Level APIs
